LUS-9378 kfilnd: Debugfs for transaction stats
While tools such as LNet selftest can be used to infer kfilnd
transaction duration, it does not provide any insight into where in
the transaction time is spent. For each kfilnd LNet network
interface, four sets of transaction stats are recorded: average time,
in nsecs, an initiator and target transaction is spent in each
transaction state with various LNet message lengths, and the average
time, in nsecs, an initiator and target transaction takes to
complete with various LNet message lengths. In addition, a reset file
is provide to reset all the stats.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>
Change-Id: I36b29757c8c3c3372737e9dc2b69dec4cb252caa
Reviewed-on: https://es-gerrit.dev.cray.com/157895
Reviewed-by: Frank Zago <fzago@cray.com>
Reviewed-by: Ian Ziemba <iziemba@cray.com>
Tested-by: Ian Ziemba <iziemba@cray.com>