LUS-9388 kfilnd: Fix immediate mrecv buf sizing
Prior to this change, immediate multi-receive buffers were sized based
on hard coded values. This prevents the ability to properly size the
immediate multi-receive buffers and can lead to messages failing due
to the target having no buffers posted.

The new approach sizes multi-receive buffers based on the number of
network credits and the hardware immediate data transfer threshold of
4096 bytes. Changing the number of network credits changes the size of
the multi-receive buffer.

In addition, a new module parameter, immediate_rx_buf_count, is
introduced to control the number of posted multi-receive buffers. 2 is
the minimal number of buffers that can be posted. On server nodes, it
may be beneficial to increase this number.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>
Change-Id: I36d94a8772157863562a264e0c254440a958c02d
Reviewed-on: https://es-gerrit.dev.cray.com/157913
Reviewed-by: Amith Abraham <aabraham@cray.com>
Reviewed-by: Ian Ziemba <iziemba@cray.com>
Tested-by: Ian Ziemba <iziemba@cray.com>