LUS-9232 test: cancel layout lock on replay deadlock
layout locks are not replayed and instead cancelled as unused, what
requires to take lov_conf_lock. the semaphore may be already taken by
cl_lock_flush() which prepares a new IO which is not be able to be
sent to MDS as it is in the recovery.

Test-Parameters: testlist=replay-dual/30
Signed-off-by: Vitaly Fertman <c17818@cray.com>
Change-Id: I54c2f4c19726be1a553b2d8cd2e7ac62dfe97066
Reviewed-on: https://es-gerrit.dev.cray.com/158090
Reviewed-by: Andriy Skulysh <c17819@cray.com>
Tested-by: Alexander Lezhoev <c17454@cray.com>