LUS-9543 lnet: Do not allow toggling discovery via lnetctl
Toggling discovery on/off at runtime via lnetctl can lead to problems
with a mounted filesystem. In particular, the state of the LNet peer
can become mismatched with ptlrpc's view of a peer and its nids.

Do not allow discovery to be toggled via lnetctl unless an
undocumented '-f' flag is provided.

Signed-off-by: Chris Horn <chris.horn@hpe.com>
Change-Id: Ia6d83c6ab8aa942f727bebae82ff74149662e8ce
Reviewed-on: https://es-gerrit.dev.cray.com/158097
Tested-by: Jenkins Build User <nssreleng@cray.com>
Reviewed-by: Petros Koutoupis <pkoutoupis@cray.com>
Reviewed-by: Alexander Boyko <c17825@cray.com>