LUS-9579 test: some test failures after the fix
Some tests are ignoreed in automation as 4 OSTs are needed.
They fail on the pool check because it is ignored after the fix
as the ost list of not the pool members is given.

Also fix the lfs setstripe util, the pool and ost list parameters
are still allowed to be given at the time.

Fixes: ("f6526365c0 LUS-9579 pool: wrong usage with ost list")
Signed-off-by: Vitaly Fertman <c17818@cray.com>
Change-Id: I6d8d537d0ccc5af47e1b0c6bceeab6eaf9930109