LUS-9931 kfilnd: Fix setting transaction nob
The kfilnd transaction number of bytes is suppose to be the LNet
payload byte count. But, the tn_nob field was not being set for all
operations in addition to being set incorrectly. The end result of
this bug is the mlength being reported in LNet reply events was
incorrect. This prevented Lustre from running.

Correctly set tn_nob to the LNet payload size. In addition, report a
mlength of zero for LNet reply events if an error has occurred during
the transaction.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>
Change-Id: Iaba794d9105827ffa3fc2d1e0be989b49fcd1ded
Reviewed-on: https://es-gerrit.dev.cray.com/158588
Reviewed-by: Amith Abraham <aabraham@cray.com>
Reviewed-by: Ian Ziemba <iziemba@cray.com>
Tested-by: Ian Ziemba <iziemba@cray.com>