LUS-9943 target: set OBD_MD_FLGRANT in read's reply
If tgt_grant_shrink() decides to not shrink grants - a client is
supposed to restore its cl_grant_avail in osc_update_grant(). In case
of read OBD_MD_FLGRANT is not set on reply's body->oa.o_valid, so
osc_update_grant() misses the cl_grant_avail update.  As result server
assumes that a client is given max number of grant while client thinks
that it is missing grants badly. That may lead to performance
degradation.

Change-Id: I0d62e15b1031fc1a40db61c718b71ae33a7050ac
Signed-off-by: Vladimir Saveliev <c17830@cray.com>
Reviewed-on: https://es-gerrit.dev.cray.com/158606
Reviewed-by: Andrew Perepechko <c17827@cray.com>
Reviewed-by: Alexander Zarochentsev <c17826@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>