LUS-9980 ldiskfs: Initial support ldiskfs on OpenSUSE 15
As of OpenSUSE 15 the primary differentiation from SUSE 15
from a lustre perspective is the kernel versioning scheme.
Where the major vendor release sp0 matches lp150, sp1 matches
lp151, and so on.

This patch enables future differentiation of minor vendor
kernel releases to be done if and when it needed.

This patch enables ldiskfs support for OpenSUSE 15.1 kernel
version 4.12.14-lp151.28.79

Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: I1cae8cdb5bd6aa45a933dd40b23a06b5e02194ed
Reviewed-on: https://es-gerrit.dev.cray.com/158647
Reviewed-by: Petros Koutoupis <pkoutoupis@cray.com>
Reviewed-by: Chris Horn <hornc@cray.com>
Tested-by: Alexander Lezhoev <c17454@cray.com>