LUS-10110 osc: Notify server if cache discard takes a long time
Discarding a large number of pages from a mapping under a
single lock can take a really long time (750GB is over 170s).
Since there is no stream of RPCs sent to the server as with
read or write to prolong the DLM lock timeout, the server
may evict the client as it does not see progress is being made.

As such send periodic "empty" RPCs to the server to show the
client is still alive and working on the pages under the lock.

For compatibility reasons the RPC is formed as a one-byte
OST_READ request with a special flag set to avoid doing
actual IO, but older servers actually do the one-byte read

Change-Id: Ib8def9ca5aae5eb16ddc877d1e4fd412df50bf64
WC-bug-id: LU-14711
Signed-off-by: Andriy Skulysh <c17819@cray.com>
Reviewed-on: https://es-gerrit.dev.cray.com/159052
Reviewed-by: Alexander Zarochentsev <c17826@cray.com>
Tested-by: Jenkins Build User <nssreleng@cray.com>
Reviewed-by: Vitaly Fertman <c17818@cray.com>