LUS-10120 osc: handle removal of NR_UNSTABLE_NFS
In Linux 5.8 the NR_UNSTABLE_NFS page counters are go.  All pages that
have been writen but are not yet safe are now counted in NR_WRITEBACK.

So change osc_page to count in NR_WRITEBACK, but if NR_UNSTABLE_NFS
still exists in the kernel, use a #define to direct the updates to
that counter.

This is a port of:

Lustre-commit: 3e5faa441266cd8dc2ee54ae140ad0129b4affa0
Lustre-change: https://review.whamcloud.com/39260

WC-bug-id: LU-13783
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Signed-off-by: Mr NeilBrown <neilb@suse.de>
Change-Id: I72004c4c76bb883799f4044aed4dbfec9a9907a4
Reviewed-on: https://es-gerrit.dev.cray.com/158889
Reviewed-by: Petros Koutoupis <pkoutoupis@cray.com>
Reviewed-by: Alexander Lezhoev <c17454@cray.com>
Tested-by: Alexander Lezhoev <c17454@cray.com>