LUS-10121 libcfs: support removal of kernel_setsockopt()
Linux 5.8 removes kernel_setsockopt() and kernel_getsockopt(), and
provides some helper functions for some accesses that are
not trivial.

This patch adds those helpers to libcfs when they are not available,
and changes (nearly) all calls to kernel_[gs]etsockopt() to
either use direct access to a helper call.

->keepalive() is not available before v4.11-rc1~94^2~43^2~14
and there is no helper function, so for SO_KEEPALIVE we
need to have #ifdef code in the C file.

TCP_BACKOFF* setting are not converted as they are not available in
any upstream kernel, so no conversion is possible.

Also include some minor style fixes and change lnet_sock_setbuf() and
lnet_sock_getbuf() to be 'void' functions.

This is a port of:

Lustre-commit: 99d9638d6c074b48f1c21c5c94d6dfe347eed3ee
Lustre-change: https://review.whamcloud.com/39259

WC-bug-id: LU-13783
Test-Parameters: trivial
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: I99aaa4dc85412c3d57b1af445bd57a65c64de9da
Reviewed-on: https://es-gerrit.dev.cray.com/158890
Reviewed-by: Andrew Perepechko <c17827@cray.com>
Reviewed-by: Petros Koutoupis <pkoutoupis@cray.com>
Tested-by: Jenkins Build User <nssreleng@cray.com>