LUS-10090 test: fix YAML support for DOM files
LFS getstripe never reports LLAPI_LAYOUT_DEFAULT for any stripe
parameter, but 0 or -1 whatever is appropriate.

LU-3285 added extra verification for the DOM parameters, precisely
the stripe count, size and offset have no sense for DOM and are
expected to be LLAPI_LAYOUT_DEFAULT. However, this brakes the yaml
support which uses getstripe output as the wanted values.

Signed-off-by: Vitaly Fertman <c17818@cray.com>
Change-Id: I7f0e722f4bcb650aac17d941b5c335caf5732235
Reviewed-on: https://es-gerrit.dev.cray.com/158898
Reviewed-by: Sergey Cheremencev <c17829@cray.com>
Reviewed-by: Alexander Boyko <alexander.boyko@hpe.com>
Tested-by: Alexander Lezhoev <c17454@cray.com>