LUS-10151 ofd: update lvb before dropping server lock
If this is a local lock on a server namespace and this was
the last reference, lock will be canceled directly instead
of staying at LRU, this might be problem that we will lose
O_DIRECT write size expanding without LVB updating.

This patch try to fix the problem by updating LVB just after
writing but before lock is dropped.

This is a port of:

Lustre-change: https://review.whamcloud.com/37790
Lustre-commit: 92d799217aeac4e1e859264d157c62e0f02b83a8

WC-bug-id: LU-13227
Fixes: 6bce536725 ("LU-4198 clio: turn on lockless for some kind of IO")
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: Idb57f49e148f81a40407d2186664a793df8183e5
Reviewed-on: https://es-gerrit.dev.cray.com/158935
Reviewed-by: Alexander Boyko <c17825@cray.com>
Reviewed-by: Petros Koutoupis <pkoutoupis@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>