LUS-10280 osc: Remove lockless truncate
Lockless truncate does not work and cannot be made to work.

Fundamentally, it has no means of ensuring consistency
across clients because it can't force them all to drop
cached data without locking.

It's been off for years - let's just get rid of it.

This is a port of:

Lustre-commit: 6335dba83995765c1ffcd7993eb8958c162913e1
Lustre-change: https://review.whamcloud.com/44204

WC-bug-id: LU-14838
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: I08aaa6cd268d4b49cf14b218a299f7246c4e950f
Reviewed-on: https://es-gerrit.dev.cray.com/159158
Reviewed-by: Alexander Boyko <c17825@cray.com>
Reviewed-by: Petros Koutoupis <pkoutoupis@cray.com>
Tested-by: Alexander Lezhoev <c17454@cray.com>