LUS-10133 kfilnd: Limit lnet_cpt_of_nid() calls
Prior to this change, lnet_cpt_of_nid() would be called by the
initiator of an LNet transaction on the critical path.
lnet_cpt_of_nid() really only needs to be called once for peer instead
of during every transaction. Fix this by using lnet_cpt_of_nid() to
determine the prefer_rx when a peer is allocated.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>
Change-Id: Ia32a91f820d9da68da42d2c22dbaf42a271ffa80
Reviewed-on: https://es-gerrit.dev.cray.com/159123
Reviewed-by: Chris Horn <hornc@cray.com>
Reviewed-by: Ian Ziemba <iziemba@cray.com>
Tested-by: Ian Ziemba <iziemba@cray.com>