LUS-9776 clio: turn on lockless for some kind of IO
We can safely turn on lockless for Direct IO
and no lock.

Direct IO will still enqueue lock in the server side,
and we could not use lockless for in the following case:

1) If group lock is held before DIO, use lockless will
make us deadlock, so we use group lock instead and trust
this to protect consistecy.

2) Direct IO might fallback to Buffer IO in some cases,
and we will restart Direct IO with normal lock holding

The main motivation for this patch is to support AIO.

This is a port from:

Lustre-commit: 6bce536725efd166d2772f13fe954f271f9c53b8
Lustre-change: https://review.whamcloud.com/8201

WC-bug-id: LU-4198
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: Id97ecb6e77e003b0d0d2a595fcf87b72be1aefd3
Reviewed-on: https://es-gerrit.dev.cray.com/158599/
Reviewed-by: Alexander Boyko <c17825@cray.com>
Reviewed-by: Petros Koutoupis <pkoutoupis@cray.com>
Tested-by: Alexander Lezhoev <c17454@cray.com>