LUS-6620 kfilnd: Handle kfabric -EAGAIN
Kfabric (and the kCXI provider) may return -EAGAIN for operations
involving the queueing of underlying hardware commands. The
expectation is the kfabric client (kfilnd) will retry the operation.

The following are the kfilnd impacted operations:
- Data transfer operations (send, tagged send, read, write)
- Post receive (tagged and untagged)
- Cancel receive (tagged and untagged)

Two kfilnd objects are impacted: transactions and immediate buffers.

To handle -EAGAIN with transactions, the transaction state machine has
been updated to only allow for one kfabric operation per transaction
state. This has resulted in the creation of two new transaction
states: TN_STATE_TAGGED_RECV_POSTED and TN_STATE_SEND_FAILED. In
addition, each transaction state will only perform non-idempotent
operations if the kfabric operation returns success. The end result is
that -EAGAIN is now propagated up to kfilnd_tn_event_handler() which
will then queue the transaction on an endpoint replay queue.

To handle -EAGAIN with immediate buffers, when the final reference is
returned for an immediate buffer and the posting of the buffer returns
-EAGAIN, the reference is returned and the immediate buffer is queued
on an endpoint replay queue for replay.

The endpoint will attempt to replay all queued operations once a timer
expires or an endpoint completion queue is drained.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>
Change-Id: Ia961200b98e810e579d44fc725f984956470909f
Reviewed-on: https://es-gerrit.dev.cray.com/159185
Reviewed-by: Ian Ziemba <iziemba@cray.com>
Tested-by: Ian Ziemba <iziemba@cray.com>
Reviewed-by: Chris Horn <hornc@cray.com>