LUS-10318 build: Disable install directive in ko2iblnd.conf
The install directive in ko2iblnd.conf is preventing the ko2iblnd
module from loading on Shasta systems. This is because we need to
load the module using the '--allow-unsupported' flag to modprobe, and
the install directive in ko2iblnd.conf is not passing that flag along
when the module is actually inserted.

Fix is to simply comment out the install directive in ko2iblnd.conf

Signed-off-by: Chris Horn <chris.horn@hpe.com>
Change-Id: I2e6f17ff6643b2b04196418d2d538ccb5d95c0d1
Reviewed-on: https://es-gerrit.dev.cray.com/159217
Reviewed-by: Shaun Tancheff <stancheff@cray.com>
Reviewed-by: Petros Koutoupis <pkoutoupis@cray.com>
Tested-by: Chris Horn <hornc@cray.com>