LUS-10270 ldiskfs: properly take inode_lock() for truncates
Originally Lustre grabbed the inode_lock() but this lead to
deadlocks as described in LU-6446 and LU-4252. The recent work
of LU-10048 changed the truncate code so that it is called
asynchronously from the main transactions. This should avoid
lock ordering issues. It should be safe to take the
inode_lock() around ldiskfs_truncate() and remove the WARN().

This is a port of:

Lustre-commit: f64e9f19f68e3983024520bd80122f12367a235a
Lustre-change: https://review.whamcloud.com/37116

WC-bug-id: LU-12977
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: I31f57bbad6954facba3ea47b918474e02d89f95a
Reviewed-on: https://es-gerrit.dev.cray.com/159252
Reviewed-by: Andriy Skulysh <c17819@cray.com>
Tested-by: Jenkins Build User <nssreleng@cray.com>
Reviewed-by: Andrew Perepechko <c17827@cray.com>