LUS-10330 kfilnd: Do not use IPv4 addressing
The usage of IPv4 addressing requires the kCXI provider to use ARP to
translate the IPv4 address into a MAC address. The assumption here is
that the MAC address would encode the RDMA physical address used which
the kCXI provider would use for data transfer source and destination
addressing. This assumption is broken if the target NIC has Ethernet
bonding configured. Depending on the bonding configuration, multiple
physical NICs may have the same MAC address. This causes the kCXI
provider to use the same RDMA physical address for independent NICs,
and this will break the RDMA protocol.

To resolve this issue, RDMA physical addressing (NIC address) is used
to build the LNet NIDs instead of an IPv4 address. When LNet is
configured, users specify a kfabric provider interface to be used for
the LNet interface. For example, kfi99(cxi0) translate to LNet NIDNET
kfi99 on local NIC index 0 using the CXI provider. As a part of the
kfabric object allocation, the true NIC address is returned. This NIC
address is used as the LNet NID addr.

The following is an example how lnetctl can be used to configure a
kfi interface.

lnetctl net add --net kfi9 --if cxi0
lnetctl net del --net kfi9 --if cxi0

This change does limit kfilnd to only be functional with the KFI CXI
provider.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>
Change-Id: I4744c029f6f4d605fd983f3e91e1eaf44c743d94
Reviewed-on: https://es-gerrit.dev.cray.com/159366
Reviewed-by: Chris Horn <hornc@cray.com>
Reviewed-by: Ian Ziemba <iziemba@cray.com>
Tested-by: Ian Ziemba <iziemba@cray.com>