LUS-10208 ptlrpc: out of order committed open replay
LUS-9298 allows committed opens to be replayed but
it doesn't enforce MDS to wait for them.
So an unlink can could be replayed before open and
the file wouldn't be placed into orphran list.

The patch allows to start recovery with commited transno.
It forces MDS to wait for replay of committed open.

Change-Id: I00022fd74363f9ab65d08eb0ad4bb037ba430132
Signed-off-by: Andriy Skulysh <c17819@cray.com>
Reviewed-on: https://es-gerrit.dev.cray.com/159273
Reviewed-by: Alexander Boyko <alexander.boyko@hpe.com>
Reviewed-by: Vitaly Fertman <c17818@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>