LUS-10351 kfilnd: Exchange kfilnd version
To support backwards compatibility between peers, peers now negotiate
which version of kfilnd should be used between them. For each new
peer, a hello request message is sent. The peer processes the hello
request message and returns the MIN of the peer kfilnd version and the
local kfilnd version. This is the negotiated kfilnd version and this
version is sent back to the initiator of the hello request message as
a hello response message.

While peers are undergoing the hello handshake, all other pending
operations between the peers are queued for retry until the handshake
is complete or the transaction deadline is met.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>
Change-Id: I5e836512d50c711eeaacd12ae1ace5933ad61376
Reviewed-on: https://es-gerrit.dev.cray.com/159369
Reviewed-by: Chris Horn <hornc@cray.com>
Reviewed-by: Ian Ziemba <iziemba@cray.com>
Tested-by: Ian Ziemba <iziemba@cray.com>