LUS-10368 kfilnd: Version control kfabric provider
Overtime, it is possible for kfabric providers to change versions. If
the kfabric provider changes a major version, this will result in now
two LNDs peers being unable to talk unless they have the same major
version.

To account for this issue, define parameters to allow admins to select
the provider version which should be used. The default provider
version major and minor values are aligned to the CXI provider.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>
Change-Id: I711117ac88ef025e8b2d397afd3917d20beec917
Reviewed-on: https://es-gerrit.dev.cray.com/159336
Reviewed-by: Steve Welch <welch@hpe.com>
Reviewed-by: Ian Ziemba <iziemba@cray.com>
Tested-by: Ian Ziemba <iziemba@cray.com>