LUS-10394 kfilnd: Move RX ctx xchg into hello msg
Instead of having RX context information (i/e prefer rx) exchanged in
the base kfilnd message header, since this is really peer
configuration information and there is a hello handshake, move the
exchange of RX context information into the hello message.

Two new fields are added to the hello message.
1. RX base: Defines the RX base the remote peer should use when
sending messages to the local peer. RX base is the same as the prefer
RX field which has been removed from the kfilnd msg header.
2. RX count: Defines the number of RX context the remote peer can
target when sending messages to the local peer. Currently, kfilnd is
only setup to allow for 1 RX context for peer. The exchange of this
information allows for extensibility to enable new features such as
the ability of the remote peer to round-robin over multiple local RX
contexts.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>
Change-Id: I20e10b64cb42deaded2ba522c347feb680bb1b1a
Reviewed-on: https://es-gerrit.dev.cray.com/159374
Reviewed-by: Chris Horn <hornc@cray.com>
Reviewed-by: Ian Ziemba <iziemba@cray.com>
Tested-by: Ian Ziemba <iziemba@cray.com>