LUS-8110 kfilnd: Support KFI_DIRECTED_RECV
KFI_DIRECTED_RECV requests that the communication endpoint (i/e RX
context) use the source address of an incoming message when matching
it with a receive buffer. If the source address field is set to
KFI_ADDR_UNSPEC instead of a valid kfi_addr_t, any source address may
match this receive buffer.

All kfilnd untagged multi-receive buffers are posted with
KFI_ADDR_UNSPEC to allow any incoming message from any peer to land
in these buffers.

All kfilnd tagged buffers are now posted with a specific source
address kfi_addr_t. This now restricts tagged buffers to a specific
peer and a specific key. This additional restriction of the tagged
buffer to a specific peer is the first step in being able to recycle
keys on kfilnd transaction timeout.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>
Change-Id: I0e43ebe57c92dbb57fa5cafba15a3fa6c6257a40
Reviewed-on: https://es-gerrit.dev.cray.com/159379
Reviewed-by: Steve Welch <welch@hpe.com>
Reviewed-by: Chris Horn <hornc@cray.com>
Reviewed-by: Ian Ziemba <iziemba@cray.com>
Tested-by: Ian Ziemba <iziemba@cray.com>