LUS-8110 kfilnd: Exchange peer session keys
The tag used for tagged buffers is expanded to 48-bits: 32-bits of
session key and 16-bits of transaction key. Each peer defines a
local session key which is sent to the remote peers during the hello
handshake. The local peer will use its local session key when posting
tagged buffers and the remote peer now has to used the remote session
key when targeting the local tagged buffer.

Now all tagged buffers are restricted to a specific session key,
transaction key, and peer. Now when the initiator marks a transaction
as timed out, the peer gets marked as down. When the peer is marked
down, no new references can be taken against this specific peer
structure meaning that the current local and remote session keys
within the down peer structure cannot be used. When a new transaction
targets this peer, a new peer structure will be allocated. This means
a new session key is allocated and the hello handshake and session key
exchanged will occur.

Now is should be safe to return the transaction keys on timeout.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>
Change-Id: I0238737ecbdfb9941ab7195b6dc4d89ef9f0558a
Reviewed-on: https://es-gerrit.dev.cray.com/159380
Reviewed-by: Steve Welch <welch@hpe.com>
Reviewed-by: Chris Horn <hornc@cray.com>
Reviewed-by: Ian Ziemba <iziemba@cray.com>
Tested-by: Ian Ziemba <iziemba@cray.com>