LUS-10282 lov: Move page index to top level
When doing readahead, we see an amazing amount of time
(~5-8%) just looking up the page index from the lov layer.

In particular, this is more than half the time spent
submitting pages:
         - 14.14% cl_io_submit_rw
            - 13.40% lov_io_submit
               - 8.24% lov_page_index

This requires several indirections, all of which can be
avoided by moving this up to the cl_page struct.

This is a port from:

Lustre-commit: 8d6d2914cf85be0508805cd5195163e1959652f4
Lustre-change: https://review.whamcloud.com/35470

WC-bug-id: LU-12535
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: I5f6f025a1ed1ff419fa7b50b28beea62300166cb
Reviewed-on: https://es-gerrit.dev.cray.com/159152
Reviewed-by: Alexander Boyko <c17825@cray.com>
Reviewed-by: Petros Koutoupis <pkoutoupis@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>