LU-13062 llite: return stripe_offset -1 in trusted.lov
If the trusted.lov xattr is copied by userspace to be restored later,
this results in PFL files always being restored onto OST0000 because
the kernel replaces the lmm_layout_gen field with 0 to avoid confusion
in userspace between the layout generation and the stripe offset (both
use the same field, one for input, one for output).

Instead of always returning 0 for the layout generation, return -1
for PFL layouts, so that restoring the xattr will direct the MDS to
select an appropriate OST.

For setxattr attempt patch prevent setting specific offsets if
PFL layout has an initialized entry what is considered as layout
copy attempt.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: Ia1af2bfcfa41cf1593aab44fe2fa792c3d254035
Signed-off-by: Mikhail Pershin <mpershin@whamcloud.com>
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/45252
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Jian Yu <yujian@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>