LUS-10604 mdc: adds client tunable to disable LSOM update
The patch adds mdc_lsom parameter to mdc, base on it state
client sends or not LSOM updates to MDT. By default LSOM update
is disabled.
lctl set_param mdc.*.mdc_lsom=[on|off]

For a configuration when LSOM is not used the patch helps
MDT with load avarage with a specific load when many threads
open/read/cose for a single file.

Test-Parameters: trivial
Signed-off-by: Alexander Boyko <alexander.boyko@hpe.com>
Change-Id: I0dbf5a150ffeebb92198e687b9cd24c1e2808bfe
Reviewed-on: https://es-gerrit.dev.cray.com/159757
Reviewed-by: Alexander Zarochentsev <c17826@cray.com>
Reviewed-by: Andrew Perepechko <c17827@cray.com>
Tested-by: Vitaly Fertman <c17818@cray.com>