LUS-5669 hsm: add hsm migration suport
Currently file migration is done by "lfs migrate". This has a couple
problems. It doesn't scale well to large files and/or large number of
files to copy. If an external tool (such as Robinhood) wants to
migrate a file, it has to spawn lfs to do the job, for every file,
which has a large overhead, or copy the file itself.

A solution to this issue is to move the copy operation into the HSM
copytool. The copytool already knows how to copy files, and Lustre has
control over these files. Migrating a file, from an application point
of view, becomes more or less the same as archiving a file. Since the
file copy is done by the copytool, this copy can be optimized in one
place.

This patch adds the ability for HSM to migrate a file within a single
Lustre filesystem. The posix copytool is enhanced to support the file
migration, and the lfs migrate command now supports that option with
--hsm and -a flags.

Signed-off-by: Ben Evans <bevans@cray.com>
WC-bug-id: LU-6081
Change-Id: Id72db5acb3f84292b081efac4657520f0054778b
Reviewed-on: https://es-gerrit.dev.cray.com/155246
Reviewed-by: Nathan Rutman <nrutman@cray.com>
Reviewed-by: Nikitas Angelinas <nangelinas@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>