LUS-10868 lod: ost list and pool name conflict
If the OST list is given on setstripe with the -o option, the pool is
unconfitionally dropped even if all the OSTs are in. Let the pool stay
in this case.

Also, if the start index given on setstripe with the -i option is out
of the pool, make it similar to the -o option - drop the pool.

Fixes: b384ea39e5 ("LU-14480 pool: wrong usage with ost list")
Signed-off-by: Vitaly Fertman <vitaly.fertman@hpe.com>
Change-Id: I718c237e273689048eb74044eea73de6c212395e
Reviewed-on: https://es-gerrit.dev.cray.com/160401
Reviewed-by: Andriy Skulysh <andriy.skulysh@hpe.com>
Tested-by: Alexander Lezhoev <alexander.lezhoev@hpe.com>
Reviewed-by: Alexander Boyko <c17825@cray.com>