LUS-11104 llite: Rework upper/lower DIO/AIO
One of the patches for LU-13799,
"Implement lower/upper aio"
(https://review.whamcloud.com/44209/) created a
complicated setup where the cl_dio_aio struct was used
both for the top level DIO or AIO and for the lower level
sub I/Os (corresponding to stripes).

This is quite complicated and hard to follow, so this
rewrites these two uses to be separate structs.  This
incidentally fixes at least one possible memory leak, but
is mostly a cleanup.

lustre-change: https://review.whamcloud.com/47187
lustre-commit: 51c18539338f1a23f3fa725470b5d36941b15c46

WC-bug-id: LU-15811
Fixes: 46ff761371 "LU-13799 Implement lower/upper aio"
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: If4fe0e8f93879700ce52dd0ba985c8f8e521c962
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/161830
Reviewed-by: Nikitas Angelinas <nangelinas@cray.com>
Reviewed-by: Vitaly Fertman <c17818@cray.com>
Tested-by: Alexander Lezhoev <alexander.lezhoev@hpe.com>