LUS-10895 ptlrpc: Revert wake_up_all on close_req completion
obd_get_mod_rpc_slot was changed in
  "LUS-9344 ptlrpc: Close request can't be sent"
which splits close_req to be added as non-exclusive waiters.

This places close_req waiters on the front of the wait queue,
and everything else is added to the end of the wait queue.

obd_put_mod_rpc_slot was changed in
  "LU-14741 obdclass: Wake up entire queue of requests on
   close completion"
which changes the wake_up side to 'wake up all' for the
close_req case.

These two changes seem to have conflicting approaches toward
the same goal.

It seems reasonable to proceed with removing the wake_up_all()
and trust the ordering of the wait_queue to do the right thing.

Fixes: ea7249b03c ("LUS-9344 ptlrpc: Close request can't be sent")
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: I5d5b38354dd2f62e3f22ee33c697f6a2caf6dd26
Reviewed-on: https://es-gerrit.dev.cray.com/160587
Tested-by: Alexander Lezhoev <alexander.lezhoev@hpe.com>
Reviewed-by: Vladimir Saveliev <c17830@cray.com>
Reviewed-by: Andriy Skulysh <c17819@cray.com>